Move http::client tests module to a standalone file
The mod.rs file was getting quite large...
